broken down into components and fixed scroll up botton
